requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
pydantic-settings==2.5.2
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
itsdangerous==2.2.0
jinja2==3.1.4
kubernetes==31.0.0
httpx==0.27.2
bcrypt==4.2.0
PyYAML==6.0.2
